# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIREST Sistema de Control de Inventarios ",
        "description": "Listado de uRI'S de la API Productos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Producto"
                ],
                "summary": "Listado de productos con paginacion 10",
                "operationId": "496dc08306524d77f18d124d1d6838aa",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Martillo"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Resistente"
                                                    },
                                                    "cantidad": {
                                                        "type": "string",
                                                        "example": "50"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Producto"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "48b59f126c849beb4d4de4ff83b7945f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Martillo"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Resistente"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": "25.50"
                                    },
                                    "cantidad": {
                                        "type": "number",
                                        "example": "50"
                                    },
                                    "almacen_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Producto creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/productos/{producto}": {
            "get": {
                "tags": [
                    "Producto"
                ],
                "summary": "Mostrar un producto específico",
                "operationId": "97c76744f3bc45c1106b78c9485f1883",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Martillo"
                                                },
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "Resistente"
                                                },
                                                "cantidad": {
                                                    "type": "string",
                                                    "example": "50"
                                                },
                                                "almacen_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-02-23T00:09:16.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-02-23T12:33:45.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Producto"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "4a85c66be7f35da356994a420d0d309a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Martillo"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Resistente"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": "25.50"
                                    },
                                    "cantidad": {
                                        "type": "number",
                                        "example": "50"
                                    },
                                    "almacen_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto modificado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Producto modificado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Producto"
                ],
                "summary": "Eliminar un producto",
                "operationId": "113f08ff5034bea8cf87748e1dff2153",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Producto eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}